Get an answer back from the content script.
We can't just return a value that pops out of executeScript() because webpack wraps our content script in a function. Thus, we can't get it to emit a statement on the top level, which is where executeScript() looks for a return value. Perhaps we could twiddle a webpack flag, but mucking with webpack tends to be a time sink.