Hide the highlighter before freezing the page.
Now we don't end up with a big highlighter polluting our captured page.

Consequently, move the highlighter twiddling into the background script. This is an aesthetic win because now everything toggles the highlighter in the same way; you don't have the devtools panel doing it ad hoc through evals. But really we were forced to do this because we needed to hide and re-show the highlighter around the page downloading, which already must be done in the background script due to permissions.

As a bonus, pages can't trip up our highlighter manipulation by monkeypatching different code onto the DOM methods we call, as they could when we were using eval().