rollup compiles it now! wu bundles properly. Add trainables file. Add an async tab-training annealer. Maybe it'll even run in parallel; we'll see what FF does.
One of the problems was that npm wasn't installing wu, despite it being a dependency of fathom-web. That was keeping rollup from bundling it. Perhaps it's that I'm using a file: URL to fathom-web. At any rate, yarn does the right thing, so I've switched to yarn.

Add json() plugin to fix "Unexpected token" errors parsing package.json files.