Fix Corpus Collector behavior in recent Firefoxes (like v65) of running past the end of the list of URLs and into `undefined`s.
This led to non-obvious misbehaviors like closing the blank.html placeholder tab prematurely.

The trouble is that we were waiting for a less specific flavor of update event on the blank.html page than we now need to. blank.html used to receive a single update event before but now receives one every time we close a tab holding a page to freeze: it gets an update event with {attention: false}.