Fix the duplicates once and for all. Fix #22 for real this time.
Turns out every page gets a bunch of onUpdated events, not just blank.html. That was firing off a second, spurious fathom:visitPage event, which was, understandably, doubling the save of the page. I suspect it was also leading to the "listener disconnected" (or words to that effect) errors, when the script slammed the tab before the second visit was done talking to it.