Add a sleep/retry loop to vector capture.
Even with all the other fixes, sometimes it claims "No message handler" or some such thing with regard to the 'vectorizeTab' message it sends to handleBackgroundScriptMessage. Though that should be a stable listener, present since the loading of the webext, it fails intermittently. Compounding my confounding, a retry fixes it.